+++ src/main/java/org/pentaho/di/git/spoon/model/IVCS.java
@@ -197,4 +197,6 @@ public interface IVCS {
   boolean merge();
 
   boolean cloneRepo( String directory, String url );
+
+  boolean rollback( String name );
 }